Allow optimizing featured image from document panel
